Rename sample app package name
