+++ velocidi-sample/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.velocidi">
+          package="com.velocidi.sampleapp">
 
     <application android:allowBackup="true"
                  android:label="@string/app_name"
